refactor: extract generated test cases view
